Fix early return bug in redisStorage.Listen
When calling XREAD with a blocking timeout and there is no new data
available within the timeout, XREAD will return a non-nil redis.Nil
error value. This is not an error condition in nacre, since we intend to
keep polling forever (or, at best, until the client data feed is
disconnected).