Bootstrap DigitalOcean/nginx/docker-compose deployments
Add deployment configuration for Nginx as a reverse proxy for the Go
appliation with catch-all error pages in case things go wrong.

Include helper scripts for bootstrapping DigitalOcean resources (one
droplet with a reserved IP assigned to it) and installing/configuring
dependencies on the droplet.

Add docker and docker-compose configurations for simplifying single-node
deployments by running Nacre and Redis side-by-side.